requirements.txt
web.py~=0.62
PyYAML~=5.3.1
requests~=2.24.0